requirements.txt
langchain>=0.1.0
langgraph>=0.0.10
groq>=0.4.1
python-dotenv>=1.0.0
pydantic>=2.0.0 
uvicorn
fastapi
openai
